backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
scipy==1.11.4
statsmodels==0.14.1
tensorflow==2.15.0
plotly==5.18.0
reportlab==4.0.7
python-dateutil==2.8.2
openpyxl==3.1.2
supabase==2.3.4
psycopg2-binary==2.9.9
python-dotenv==1.0.0

